Remove hook
